requirements.txt
berserk==0.13.2
python-chess==1.999
requests==2.32.3